% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemartix.R
\name{codematrix}
\alias{codematrix}
\alias{codematrix.matrix}
\alias{codematrix.numeric}
\alias{is.codematrix}
\title{Construct EQ5D code martix}
\usage{
codematrix(x, const, ...)

\method{codematrix}{matrix}(x, const, ...)

\method{codematrix}{numeric}(x, y = NULL, z = NULL, ...)

is.codematrix(x)
}
\arguments{
\item{x}{object}

\item{const}{constant values (numeric of length 3)}

\item{y, z}{additional numeric vectors if \code{x} is numeric of length 5}
}
\value{
Object of class "codematrix", a \code{5 * 3} numerical matrix
with rows corresponding to EQ5D-dimensions (+ constant),
columns correcponding to response levels and individual values
corresponding to translation weights.
}
\description{
Construct EQ5D code martix
}
\examples{

# Swedish code martix based on individual vectors for each response level
uk <- codematrix(
        rep(0, 6),
        c(.069, .104, .036, .123, .071, .081),
        c(.314, .214, .094, ..386, .237, .269))

}

