% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq2index.R
\name{eq2index.character}
\alias{eq2index.character}
\title{Calculate EQ5D-index from individual EQ5D-levels}
\usage{
\method{eq2index}{character}(x, ...)
}
\arguments{
\item{x}{EQ5D-data, see examples}

\item{...}{arguments pased between methods}

\item{cm}{codematrix, either as a a \code{\link{codematrix}} object
or a character name of such object}
}
\value{
numeric index vector
}
\description{
Calculate EQ5D-index from individual EQ5D-levels
}
\examples{
# nidividual numeric vector
eq2index(c(1, 2, 3, 2, 1))
}

