% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir}
\alias{ir}
\title{Improvement Ratio}
\usage{
ir(pre, post, best, alpha = 0.05)
}
\arguments{
\item{pre, post}{pre- and post-treatment PROM-values}

\item{best}{best (desired) post-treatment PROM value}

\item{alpha}{significance level for confidence interval}
}
\value{
object of class "ir"
}
\description{
Improvement Ratio
}
