% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir.test}
\alias{ir.test}
\title{Test if Improvemant Ratios from different groups are equal}
\usage{
ir.test(x, y)
}
\arguments{
\item{x, y}{output from \code{\link{ir}}}
}
\value{
Numeric vector with estimated Z-value and associated p-value.
}
\description{
Test if Improvemant Ratios from different groups are equal
}
