% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir.trendtest}
\alias{ir.trendtest}
\title{Trend test of Improvement Ratio}
\usage{
ir.trendtest(x, sds, score = seq_along(x))
}
\arguments{
\item{x}{output from \code{\link{ir}}}

\item{score}{}
}
\value{
Object of class "htest"
}
\description{
Trend test of Improvement Ratio
}
