% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemartix.R
\name{codematrix}
\alias{codematrix}
\alias{codematrix.matrix}
\alias{codematrix.numeric}
\alias{is.codematrix}
\title{Construct EQ5D code martix}
\usage{
codematrix(x, ...)

\method{codematrix}{matrix}(x, const, ...)

\method{codematrix}{numeric}(x, y = NULL, z = NULL, ...)

is.codematrix(x)
}
\arguments{
\item{x}{object}

\item{...}{pass arguments to other methods}

\item{const}{constant values (numeric of length 3)}

\item{y, z}{additional numeric vectors if \code{x} is numeric of length 5}
}
\value{
Object of class "codematrix", a \code{5 * 3} numerical matrix
with rows corresponding to EQ5D-dimensions (+ constant),
columns correcponding to response levels and individual values
corresponding to translation weights.
}
\description{
Construct EQ5D code martix
}
\examples{

# British
uk <-
 codematrix(
   rep(0, 5),
   c(.069, .104, .036, .123, .071),
   c(.314, .214, .094, .386, .236),
   const = c(1, .081, .35)
 )

# Swedish
se <-
 codematrix(
   rep(0, 5),
   c(.0666, .0276, .1012, .0345, .0552),
   c(.1247, .0276, .1355, .0904, .2077),
   const = c(.9694, 0, .0433)
 )
}
